requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
idna==3.4
joblib==1.2.0
kafka-python==2.0.2
nltk==3.8.1
numpy==1.24.2
praw==7.7.0
prawcore==2.3.0
py4j==0.10.9.7
pyspark==3.4.0
regex==2023.3.23
requests==2.28.2
tqdm==4.65.0
update-checker==0.18.0
urllib3==1.26.15
websocket-client==1.5.1
